feat: move sta2rest to fastapi folder
